% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/textAnalysis.r
\name{doMetrics}
\alias{doMetrics}
\title{Input unmerged corpus and run metrics on it}
\usage{
doMetrics(docs, properCorpus, settings)
}
\arguments{
\item{docs}{unmerged corpus}

\item{properCorpus}{unmerged corpus of proper words (or could be any other filtering of original docs really)}

\item{settings}{list, must include helperFiles$buzzFile1, helperFiles$buzzFile2 - links to buzzword lists, and topicCount, topicIter - values for supervised topic modelling}
}
\value{
list containing metric results, and document corpora along with supervised topic models
}
\description{
Run readability, length, topic model metrics on documents
}
\examples{
\dontrun{
doMetrics (corpus, properCorpus, list(helperFiles=list(buzzFile1="C:/Martin/Data/Impact/buzz.txt", buzzFile2="C:/Martin/Data/Impact/busibuzz.txt"), topicCount=10, topicIter=30))
}
}

