% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/textAnalysis.r
\name{drawWeightGPAPlot}
\alias{drawWeightGPAPlot}
\title{Draw scatterplot of one variable against impact scores
Draw a scatterplot of a given variable vector against a vector of impact scores to eyeball for patterns}
\usage{
drawWeightGPAPlot(impactScores, weightScores, colScale, xlab = "WEIGHT",
  ylab = "IMPACT GPA")
}
\arguments{
\item{impactScores}{vector of doubles, impact score per institution/entry}

\item{weightScores}{a vector of doubles to pair off and compare with the impact scores}

\item{colScale}{colour scale for plot or a set value}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}
}
\value{
the plot object (i think)
}
\description{
Draw scatterplot of one variable against impact scores
Draw a scatterplot of a given variable vector against a vector of impact scores to eyeball for patterns
}
\examples{
drawWeightGPAPlot (c(1.0, 1.1, 1.2), c(2.0, 2.4, 3.0), "#444444")
}

