% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/textAnalysis.r
\name{fromTheTop}
\alias{fromTheTop}
\title{A Text Loading Function}
\usage{
fromTheTop(mySettings = list())
}
\value{
list containing metric results, and document corpora along with supervised topic models
}
\description{
This function chews a load of pdfs in a fixed directory
}
\keyword{impact}
\keyword{ref}
\keyword{text}

