% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_institutions.R
\name{get_institutions}
\alias{get_institutions}
\title{List Institutions}
\usage{
get_institutions()
}
\value{
Returns a data_frame (from the \code{tibble} package).
}
\description{
This function uses the \code{ListInstitutions} method from the database API
and returns a list of institutions and associated metadata, including the
UKPRN.
}
\examples{
\dontrun{
get_institutions()
}

}

