% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tag_types.R
\name{get_tag_types}
\alias{get_tag_types}
\title{List Tag Types}
\usage{
get_tag_types()
}
\value{
Returns a data_frame (from the \code{tibble} package).
}
\description{
This function uses the \code{ListTagTypes} method from the database
API and returns a list of Tag Types and associated metadata.
}
\examples{
\dontrun{
get_tag_types()
}

}

