% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tag_values.R
\name{get_tag_values}
\alias{get_tag_values}
\title{List Tag Values}
\usage{
get_tag_values(tag_type)
}
\value{
Returns a data_frame (from the \code{tibble} package).
}
\description{
This function uses the \code{ListTagValues} method from the database
API and returns a list of Tag Values for the supplied tag type.
}
\examples{
\donttest{
get_tag_values(3) # Tag Type 3 (Subject)
}

}

