% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_units_of_assessment.R
\name{get_units_of_assessment}
\alias{get_units_of_assessment}
\title{List Units of Assessment}
\usage{
get_units_of_assessment()
}
\value{
Returns a data_frame (from the \code{tibble} package).
}
\description{
This function uses the \code{ListUnitsOfAssessment} method from the database
API and returns a list of Units of Assessment and associated metadata.
}
\examples{
\dontrun{
get_units_of_assessment()
}

}

