% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/textAnalysis.r
\name{impactScoreReader}
\alias{impactScoreReader}
\title{Read in impact scores from file}
\usage{
impactScoreReader(settings)
}
\arguments{
\item{settings,}{list containing cname: path to csv file, uoa : the unit of assessment, profile: the profile type to score}
}
\value{
a data.table of scores per university
}
\description{
This function reads the ref2014 csv file and extracts scores given a uoa and profile type
}
\examples{
impactScoreReader (cname="C:/Martin/Data/Impact/REF submission documents/Templates", uoa=19, profile="Impact")
}

