% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/textAnalysis.r
\name{loadDocs}
\alias{loadDocs}
\title{A Multiple PDF Document Loading Function}
\usage{

  loadDocs(dirName = "C:/Martin/Data/Impact/REF submission documents/Case Studies",
  isRecursive = TRUE)
}
\arguments{
\item{dirName}{directory PDFs are located in}

\item{isRecursive}{flag whether sub-directories should be searched}
}
\value{
TM document corpus
}
\description{
This function chews a load of pdfs in a given directory
}
\examples{
loadDocs (dirName="C:/Martin/Data/Impact/REF submission documents/Templates", isRecursive=FALSE)
}

