% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/avgjac.r
\name{makeLDAVisJSON}
\alias{makeLDAVisJSON}
\title{Produce visualisation of LDA topic corpus using LDAvis package}
\usage{
makeLDAVisJSON(corpus, slda, alpha = 1, eta = 0.1, dir = "vis")
}
\arguments{
\item{corpus}{TM document corpus}

\item{slda}{supervised LDA topic models}
}
\value{
alpha optional SLDA variable, 1.0 if docs contain diverse material which the merged ones do

eta an even less explanatory optional SLDA variable, never found a plain words explanation for it yet. Best leave to default.

dir the directory name in which to place generated web pages. optional, default is "vis"
}
\description{
Takes in corpus and supervised topic model and produces interactive web page with topics marked on it
}
\examples{
\dontrun{
makeLDAVisJSON (corpus, topicModel, dir="vis")
}
}

