% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/textAnalysis.r
\name{makeProperWordCorpus}
\alias{makeProperWordCorpus}
\title{Produce corpus of proper names from original corpus}
\usage{
makeProperWordCorpus(originalCorpus, settings)
}
\arguments{
\item{originalCorpus}{TM document corpus}

\item{settings}{list of boolean flags to control what is parsed (people, places, orgs, date). keepSpaces decides concatenation of multi-word terms.}
}
\value{
TM document corpus containing just proper names
}
\description{
Takes in corpus and flags to decide which entity types to parse for
}
\examples{
makeProperWordCorpus (corpus, list(keepSpaces=TRUE, people=TRUE, places=TRUE))
}

