% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/textAnalysis.r
\name{mergeByInstitution}
\alias{mergeByInstitution}
\title{# Merge documents with same metadata information (i.e name) into single documents}
\usage{
mergeByInstitution(docs, metaProp = "name")
}
\arguments{
\item{docs}{unmerged corpus}

\item{field}{in document meta information to merge on. Default is name.}
}
\value{
merged docs by uni name, with extra meta field docCount indicated number of docs merged per institution
}
\description{
Preserve content sectioning, copies over uni names and impact scores
}
\examples{
\dontrun{
mergeByInstitution (corpus)
}
}

