% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/textAnalysis.r
\name{primeCaseStudyDocs}
\alias{primeCaseStudyDocs}
\title{Corpus quality control and metadata attachment}
\usage{
primeCaseStudyDocs(docs, settings)
}
\arguments{
\item{docs}{TM document corpus}

\item{settings}{list, must include three properties: boilerPlate, sectionHeadings and impactSettings}
}
\value{
altered TM document corpus, each doc now contains a set number of sections
}
\description{
Takes in original case study corpus, attaches metadata, sections the content, filters boilerplate
and removes documents that fail some quality metric (empty sections)
Settings list must include three properties:
1) boilerPlate = array of boilerplate strings to strip from docs
2) sectionHeadings = ordered array of sectionHeadings used to split documents up
3) impactSettings = list that describes where impact scores are sourced from, cname=filepath, uoa=unit of assessment number, profile=assessment type (i.e "impact")
}

