% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/textAnalysis.r
\name{properNameExtraction}
\alias{properNameExtraction}
\title{Extract proper names from corpus}
\usage{
properNameExtraction(docs, settings)
}
\arguments{
\item{docs}{TM document corpus}

\item{settings}{list of boolean flags, set these to true if entities wanted: people, places, orgs, money, date, percentage}
}
\value{
list of proper names per document
}
\description{
Takes in corpus and flags to decide which entity types to parse for
}

