% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/textAnalysis.r
\docType{data}
\name{ref2014}
\alias{ref2014}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 53940 rows and 18 variables:
\describe{
  \item{Institution.code..UKPRN.}{uni code}
  \item{Institution.name}{university name}
}}
\source{
ref2014 website
}
\usage{
ref2014
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}

