% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_get.R
\name{ref_get}
\alias{ref_get}
\title{Call the REF Impact Case Studies API}
\usage{
ref_get(api_method, tag_type = NULL, query = NULL)
}
\arguments{
\item{api_method}{text string specifying the API method you wish to call.}

\item{tag_type}{integer, the ID of the tag type you wish to retrieve (see
below)}

\item{query}{list, search parameters for use with the case_studies method
(see below)}
}
\value{
Returns a tibble (from the \code{tibble} package).
}
\description{
This function calls the REF Impact Case Studies API.
}
\examples{
institutions <- ref_get("institutions")
units_of_assessment <- ref_get("uoa")
tag_types <- ref_get("tag_types")
tag_type_5 <- ref_get("tag_values", 5L)

}

