% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_institutions.R
\name{ref_institutions}
\alias{ref_institutions}
\title{List Institutions}
\usage{
ref_institutions()
}
\value{
Returns a tibble (from the \code{tibble} package).
}
\description{
This function uses the \code{ListInstitutions} method from the database API
and returns a list of institutions and associated metadata, including the
UKPRN.
}
\examples{
\donttest{
ref_institutions()
}

}

