% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/textAnalysis.r
\docType{data}
\name{ref2014}
\alias{ref2014}
\title{REF 2014 data}
\format{A data frame with 7644 rows and 18 variables:
\describe{
  \item{Institution.code..UKPRN.}{uni code}
  \item{Institution.name}{university name}
  \item{Institution.sort.order}{numerical sort order}
  \item{Main.panel}{assessment panel identifier}
  \item{Unit.of.assessment.number}{UOA code}
  \item{Unit.of.assessment.name}{UOA name}
  \item{Multiple.submission.letter}{identifier for multiple entries from one institution}
  \item{Multiple.submission.name}{department within university}
  \item{Joint.submission}{submission between universities}
  \item{Profile}{Aspect of REF}
  \item{FTE.Category.A.staff.submitted}{number of staff}
  \item{X4.}{4 star rating}
  \item{X3.}{3 star rating}
  \item{X2.}{2 star rating}
  \item{X1.}{1 star rating}
  \item{unclassified}{numerical sort order}
  \item{Area}{geographic area code within UK}
  \item{Post92}{new university flag}
}}
\source{
ref2014 website
}
\usage{
ref2014
}
\description{
A dataset containing REF ratings for all UK institutions
across multiple aspects and units of assessment
}
\keyword{datasets}

